cloud-functions/requirements.txt
# Google Cloud Function Requirements
google-cloud-functions==1.13.3
flask==2.3.2
requests==2.31.0
functions-framework==3.4.0
